pyproject.toml
[project]
name = "shorter-url"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "certifi>=2026.2.25",
    "fastapi>=0.135.1",
    "pybase62>=1.0.0",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
